+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Unreleased
 
 ### Added
+- Allow to exclude symlinked packages automatically
 - Compatibility with 2023.2 (eap)
 
 ## 1.18.0 - 2023-06-12